PlayerStackView 분리
